requirements.txt
Flask
Flask-PyMongo
pandas
joblib
pytesseract
pdf2image
Pillow
gunicorn
python-dotenv
scikit-learn
requests